backend/requirement.txt
amqp==5.3.1
aniso8601==10.0.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
libpass==1.9.1.post0
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
libpass==1.9.1.post0
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
WTForms==3.2.1